+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 - Maps for roles, users and account
 - Default roles to assume for EKS
+- Flag enable_default_roles to enable creation of default role; default false 
 
 ## [2.1.1] 2019-06-07
 ### Changed